+++ pkg/services/tree/container.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/hashicorp/golang-lru/simplelru"
+	"github.com/hashicorp/golang-lru/v2/simplelru"
 	"github.com/nspcc-dev/neofs-node/pkg/core/container"
 	"github.com/nspcc-dev/neofs-node/pkg/services/object_manager/placement"
 	cidSDK "github.com/nspcc-dev/neofs-sdk-go/container/id"
@@ -16,11 +16,11 @@ import (
 type containerCache struct {
 	sync.Mutex
 	nm  *netmapSDK.NetMap
-	lru *simplelru.LRU
+	lru *simplelru.LRU[string, containerCacheItem]
 }
 
 func (c *containerCache) init(size int) {
-	c.lru, _ = simplelru.NewLRU(size, nil) // no error, size is positive
+	c.lru, _ = simplelru.NewLRU[string, containerCacheItem](size, nil) // no error, size is positive
 }
 
 type containerCacheItem struct {
@@ -48,8 +48,7 @@ func (s *Service) getContainerNodes(cid cidSDK.ID) ([]netmapSDK.NodeInfo, int, e
 	s.containerCache.Lock()
 	if s.containerCache.nm != nm {
 		s.containerCache.lru.Purge()
-	} else if v, ok := s.containerCache.lru.Get(cidStr); ok {
-		item := v.(containerCacheItem)
+	} else if item, ok := s.containerCache.lru.Get(cidStr); ok {
 		if item.cnr == cnr {
 			s.containerCache.Unlock()
 			return item.nodes, item.local, nil