+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog for NeoFS Node
 ### Updated
 - Update minimal supported Go version up to v1.18 (#2340)
 - tzhash library to 1.7.0 (#2348)
+- `github.com/hashicorp/golang-lru` to `v2.0.2`
 
 ### Updating from v0.36.1